automatically backed up by learn

+++ db/migrate/002_create_costume_stores.rb
@@ -5,7 +5,7 @@ class CreateCostumeStores < ActiveRecord::Migration[5.1]
     create_table :costume_stores do |t|
       t.string :name
       t.string :location
-      t.integer :custume_inventory
+      t.float :custume_inventory
       t.integer :number_of_employees
       t.boolean :whether
       t.timestamps
